Complete refactor of the Twig view for straightforward overriding
The base class for Twig view is set up to be usable with sane defaults without extension, and to be easily extended to allow redefining parameters for Twig and defining any additional functions or filters without requiring modifying the base class in any way.

The unused import statements have been removed as they will not be used in this base class, only in extensions.