Capture client's accept language header contents
As localisation is an open-ended problem that can be handled in any number of ways by an arbitrary application, the framework cannot make any assumptions about how to process or decide anything about it. As a result, this change only facilitates passing along the accept language header's contents down to the controller level for custom processing to be defined by an application however.

In the future a usable controller may be added to the framework which an application may elect to utilise that provides some sort of fundamental processing of this header's contents e.g. returning a ranked sequence of locales being requested, but at this revision only a handed down copy of the value is provided.

This commit also removes an overlooked class import which became obsolete.