Post files abstraction
The standard super-global for POST-delivered files is wrapped and transformed into an array of basic data structures representing the uploaded files. This array is automatically set as an input for every controller for ready access.

This does not only abstract the language's built-in representation of this data, but also addresses its non-intuitive structure and provides enforced protection with ensuring everything is a proper uploaded file.