Session requirement enforcement
The framework provides an option for requiring sessions for any given route or group as a convenience, when the outcome is only to redirect the user to a programmable route if the user is not authenticated.

This feature remains optional so that an application can ignore its effects and define its own method of handling authentication enforcement through usage of controllers.