Replace DB credentials with a simple PDO wrapper that allows lazy loading
Previously wrapping all of PDO's constructor was attempted and was a cumbersome effort that lead to reimplementing existing logic built into the language's library. This replacement is a straightforward substitute that eliminates that redundancy, and will not require further maintenance if new PDO drivers are added to the language in the future.