Add memory store class for memcached
Memcached is the other significant alternative to Redis for memory stores which is now supported. Configurability of the memory store connection settings will need to be added later. For now, a typical default is defined.

Extending the memory store class and overriding the constructor is a current solution to custom connection settings.