Add Spout view
Added a binary view for the Spout spreadsheet export library with the export wrapper and formatter model classes to accompany it. Concrete views such as this are always optional and do not infringe on the open-ended design of the framework's view layer.

Application developers may choose to disregard this built-in support, and supply their own binary view definition. This will never be restricted.