Allow configuration to reroute root index requests
Previously it was necessary to add a rewrite directive to the web server daemon for the framework to process a root index request, since the framework would always consider such a request as malformed. This change introduces the ability to capture this request and process it accordingly by permitting the configuration to map an ordinary route as the handler for such requests.