Output mapping
Output mapping allows a route's collection of outputs to be remapped based on a list of aliases provided by the output map. This allows greater portability and reuse of various controllers with potential templates by keeping variable names consistent in templates regardless of the naming conventions outputs bear in existing controllers. Not only can existing keys for values be aliased, but keys not reiterated in the map are suppressed, mostly valuable for JSON views as explained below.

This feature also addresses the security concern of leaking data. Previously JSON responses would uncontrollably echo everything generated by controllers for output, meaning all data would be passed to the client directly. With an output map, the data released to the client via JSON is fully controlled, and sensitive or unnecessary data can be withheld.

The data leak issue was not applicable to other view modes due to specific rendering being performed for those, which inherently provide the ability to control what is passed through for the final output.