Allow null for accept language
It is possible for an HTTP request to completely omit the accept language header, which would be typical with tools such as wget and curl. In such cases the environment variable is actually null, not empty string, which was not expected by the framework's handling of this value. Consequently a null would raise an internal script error which would direct execution flow to the kernel's fallback exception handler unintentionally, yielding HTTP 500 failures when normally the request was otherwise appropriate.

Safely permitting null for the internal copy of accept language resolves this problem.