Refactor kernel exception handling
This refactor addresses two issues with the kernel's exception handling/catch block. First, the redundancy of the actions across the multiple catch block was high with very few differences, and consequently has been eliminated. Second, and more importantly, the possibility of an uncaught exception escaping the catch block has been addressed.

Such a failure would only occur with a misconfiguration, but gracefully handling the occurrence will reduce the impact of such a situation, and properly logs the incident should it actually occur.