Facilitate memory store connection settings by configuration
Memory stores are no longer restricted to default connection settings. The configuration is now obligated to define the logic that instantiates the client object for the actual memory store and return the wrapper for it accordingly which is compatible for usage within the kernel.

This completely open-ended approach is necessary due to memory stores having wildly varying configuration options.