Add Controller Slice class
The foundation of version 2 is the new controller slices. In version 1,
a sequence of controllers could be broken by emitting failure
intentionally and following specific behaviour for failure. However,
this was limited to what behaviours could be defined. For longer
controller sequences, it may be necessary to abort in different ways
depending on which controller in the sequence failed, or perhaps has a
variation in successful return that warrants an early abort.

This limitation, along with other unintended constraints such as the
missing flexibility of aliasing controller I/O and properly protecting
the view layer's variable visibility, prompted the creation of this
encapsulating structure around controllers. Slices will serve as an
adapter between controllers and the framework itself, dictating how to
respond or continue during controller execution.