Add Exit State class
Exit State is the new object that slices will map that establishes the
association between how controllers return and how the framework will
respond with respect to those states. This constitutes a significant
breaking change from version 1. To facilitate the more advanced
configuration and routing capabilities for controllers and their
encapsulating slices, it is necessary to use a more flexible return
value to signal the framework how to behave at the conclusion of a
controller's execution.

Inherently this also solves design limitations perceived in the previous
version that made controller chains too rigid, particularly with regard
to forking behaviour depending on which controller in the sequence is
the one that raised a failure. This commit brings the class definition
into the project and subsequence changes will incorporate it.