Upgrade kernel to handle new exit state signals
Main methods of controllers will now return a proper return code rather
than a primitive Boolean result of success, and the kernel is updated to
incorporate and proliferate this approach as needed. A slight amount of
refactoring was done to separate the logic to handle a redirection
signal outside of the routine that processes the rendering for a view.