Support zero-slice routes correctly by invoking the default exit state
A subtle issue exists with a route that uses no slices: the exit map
would normally be overlooked since there are no controllers to execute.
This change checks for this situation and simply implies the first
defined exit state map is presumed for engagement to determine the
response of such a route.

The absence of an exit state map will raise an unrecoverable failure,
but superfluous definitions of states are redundant and ignored.