Proper store I/O mapping with alias support
Handling of store input and output maps has been refined to channel keys
and values correctly. Additionally an intuitive solution for processing
optional aliasing in either map is included. This allows configurations
to avoid redundantly repeating names in map definitions when aliasing is
not desired, which is the common case, but still allow straightforward
inline alias declarations.