Session output map processing must open and close the session
Attempting to write to the session store during store output map
processing without properly ensuring the session was opened first led to
a memory store exception reporting the failure to write to a session of
an inactive state. This change ensures the opening of the session prior
to processing the mappings, and closes the session afterwards cleanly to
release the lock.