Extend MemoryStore with volatile store
The volatile store is another form of access to the existing local store
whose only behavioural difference is that accessing a key in this way
causes the value to be subsequently destroyed automatically.

This allows one-time usage without further micromanagement of the memory
store by controllers or route definitions to achieve the same effect.
Persisting request data and driving PRG patterns are the most prominent
examples of processing workflows that are straightforwardly enabled by
using this new store.