Procure the protocol & version from the environment
Replaced the old sane default of HTTP/1.1 with the environment variable
declaring the protocol and version, and defining that value accordingly
in the headers object. This more appropriate solution was motivated by
HTTP/2.0, although interestingly enough, there didn't appear to be
incorrect headers being sent before this patch under HTTP/2.