Route rewriting support
The concept of rewriting adopts the principle of sending the client a
redirect in order to complete one execution sequence and lead
immediately to another by offering the option of similar behaviour
without terminating execution with a rewrite. In other words, a route
may now conclude its own slice sequence and divert directly into the
start of any other route as a valid outcome in exit state mapping, as
though they were concatenated together explicitly.

Routes may bear lengthy or otherwise complicated configurations and some
applicaton designs may seek to reuse those routes with slightly varying
entry points, as a leading example. Without rewriting, true redirects
with some memory management gymnastics would be necessary, or
substantial duplication of routing group programming to achieve the
effect without internally rewriting the route. Another benefit from this
option is that because the context does not change, inputs and request
data persist, allowing intuitive combinations with minimal verbosity.

The naming convention attributed to this feature comes from the concept
that the original request is being rewritten into another, representing
the runtime concatenation of two or more routes. As it is an exit map
driven event, it is effectively conditional to rewrite, analogous to
responding with a redirection.