Complete implementation of permissions for restricting route access
The original version of the framework had incomplete logic for
permissions that was dormant and inaccessible. This was rectified, and
permissions are now properly implemented for actual usage.

Accessing a route protected by permission token without possessing the
corresponding token results in an automatic response with unsuccessful
response code. Successful clearance leads to ordinary execution of
slices as though no permissions were in effect.

The configuration class is used to allow algorithmic definition for
retrieving a set of permissions from a session. It also presents the
permission set structure as an interface in essence, allowing more
complex implementations to replace the basic hash table-driven solution
provided without altering kernel code in appropriate polymorphic
fashion.