Ignore unset keys when processing the output map
Controllers are not obligated to define values for every mapped key. If
a controller must define a value, or destroy a value, it can emit
something, or null, respectively.

By ignoring undefined keys the kernel will not raise notices in the PHP
engine, and, more importantly, controllers do not need to have
tedious extraneous statements for key/value assignment.