Process the slice output map
The basis for the slice output map was set but the processing during
controller execution was never defined. Adhering to precedent, only
variables actually emitted by the controller will be considered when the
map is being iterated.

The default behaviour is preserved, which is to merge all emitted
outputs from the controller.